backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis==5.0.1

# Celery for background tasks
celery[redis]==5.3.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
cryptography==42.0.2

# Pydantic settings
pydantic[email]==2.6.1
pydantic-settings==2.1.0

# AWS S3
boto3==1.34.34

# HTTP client for external APIs
httpx==0.26.0
requests==2.31.0

# AI Providers
openai==1.51.0
anthropic==0.39.0
google-generativeai==0.8.3

# Logging and monitoring
python-json-logger==2.0.7

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.26.0
